Refactored Io code
