+++ boards/BMS/Inc/Io/configs/Io_LTC6813Configs.h
@@ -1,6 +1,5 @@
 #pragma once
 
 #define NUM_OF_LTC6813 2U
-#define NUM_OF_CELLS_PER_LTC6813 18U
 #define NUM_OF_CELLS_READ_PER_LTC6813 16U
-#define LTC6813_SPI_DELAY_MS 2U
+#define SPI_TIMEOUT_MS_LTC6813 2U
